chore: extract active project hooks to a separate file
