+++ packages/frontend/src/components/SettingsUsageAnalytics/SettingsUsageAnalytics.styles.ts
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const CardContent = styled.div`
+    display: flex;
+    gap: 10px;
+    height: 24px;
+`;
