+++ packages/frontend/src/hooks/useProject.ts
@@ -6,15 +6,9 @@ import {
     UpdateProject,
 } from '@lightdash/common';
 import { useMutation, useQuery, useQueryClient } from 'react-query';
-import { useParams } from 'react-router-dom';
 import { lightdashApi } from '../api';
 import { useActiveJob } from '../providers/ActiveJobProvider';
 import useToaster from './toaster/useToaster';
-import {
-    useActiveProject,
-    useDefaultProject,
-    useProjects,
-} from './useProjects';
 import useQueryError from './useQueryError';
 
 const createProject = async (data: CreateProject) =>
@@ -95,32 +89,4 @@ export const useCreateMutation = () => {
             },
         },
     );
-};
-
-export const useActiveProjectUuid = () => {
-    const params = useParams<{ projectUuid?: string }>();
-    const { data: projects, isLoading: isLoadingProjects } = useProjects();
-    const { data: defaultProject, isLoading: isLoadingDefaultProject } =
-        useDefaultProject();
-    const { data: lastProjectUuid, isLoading: isLoadingLastProject } =
-        useActiveProject();
-
-    if (isLoadingProjects || isLoadingDefaultProject || isLoadingLastProject) {
-        return {
-            isLoading: true,
-            activeProjectUuid: undefined,
-        };
-    }
-
-    const lastProject = projects?.find(
-        (project) => project.projectUuid === lastProjectUuid,
-    );
-
-    return {
-        isLoading: false,
-        activeProjectUuid:
-            params.projectUuid ||
-            lastProject?.projectUuid ||
-            defaultProject?.projectUuid,
-    };
 };
