+++ packages/frontend/src/pages/Projects.tsx
@@ -2,7 +2,7 @@ import { FC } from 'react';
 import { Redirect } from 'react-router-dom';
 import ErrorState from '../components/common/ErrorState';
 import PageSpinner from '../components/PageSpinner';
-import { useActiveProjectUuid } from '../hooks/useProject';
+import { useActiveProjectUuid } from '../hooks/useActiveProject';
 import { useProjects } from '../hooks/useProjects';
 
 const Projects: FC = () => {
